refactor(landing): add landing screen with animated sections
This commit introduces a new LandingScreen composable that serves as the
main entry point for the application. It features various animated
sections, including a header, service categories, and a footer call to
action. The design incorporates a modern aesthetic with smooth
transitions and interactive elements to enhance user engagement.

Additionally, the MainViewController is updated to initialize Koin
dependency injection, ensuring that the application is set up correctly
for dependency management on iOS.
