+++ androidApp/src/main/kotlin/com/example/seviya/MainActivity.kt
@@ -6,6 +6,7 @@ import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.tooling.preview.Preview
+import com.example.seviya.app.App
 import com.google.firebase.Firebase
 import com.google.firebase.initialize
