+++ composeApp/src/androidMain/kotlin/com/example/seviya/UI/ClientPaymentUploadScreen.kt
@@ -22,9 +22,9 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.example.seviya.theme.BrandBlue
-import com.example.seviya.theme.BrandRed
-import com.example.seviya.theme.White
+import com.example.seviya.core.designsystem.theme.BrandBlue
+import com.example.seviya.core.designsystem.theme.BrandRed
+import com.example.seviya.core.designsystem.theme.White
 import com.example.shared.presentation.ClientPaymentUpload.ClientPaymentUploadViewModel
 import io.kamel.image.KamelImage
 import io.kamel.image.asyncPainterResource
