+++ composeApp/src/androidMain/kotlin/com/example/seviya/UI/WorkerDailyAppointmentsScreen.kt
@@ -62,9 +62,9 @@ import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.example.seviya.theme.BrandBlue
-import com.example.seviya.theme.BrandRed
-import com.example.seviya.theme.White
+import com.example.seviya.core.designsystem.theme.BrandBlue
+import com.example.seviya.core.designsystem.theme.BrandRed
+import com.example.seviya.core.designsystem.theme.White
 import com.example.shared.domain.entity.Appointment
 import com.example.shared.presentation.workerDailyAppointments.DailyView
 import com.example.shared.presentation.workerDailyAppointments.WorkerDailyAppointmentsViewModel
