+++ composeApp/src/commonMain/kotlin/com/example/seviya/UI/ClientAppointmentDetailScreen.kt
@@ -37,28 +37,28 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
-import com.example.seviya.components.ClientBottomBar
-import com.example.seviya.components.ClientTab
-import com.example.seviya.theme.AppBackground
-import com.example.seviya.theme.BlueGrayText
-import com.example.seviya.theme.BorderSoft
-import com.example.seviya.theme.BrandBlue
-import com.example.seviya.theme.BrandRed
-import com.example.seviya.theme.SoftSurface
-import com.example.seviya.theme.StatusCancelledBackground
-import com.example.seviya.theme.StatusCancelledText
-import com.example.seviya.theme.StatusCompletedBackground
-import com.example.seviya.theme.StatusCompletedText
-import com.example.seviya.theme.StatusConfirmedBackground
-import com.example.seviya.theme.StatusConfirmedText
-import com.example.seviya.theme.StatusInProgressBackground
-import com.example.seviya.theme.StatusInProgressText
-import com.example.seviya.theme.StatusPendingBackground
-import com.example.seviya.theme.StatusPendingText
-import com.example.seviya.theme.TextBluePrimary
-import com.example.seviya.theme.TextOnBlueSubtitle
-import com.example.seviya.theme.TextSecondary
-import com.example.seviya.theme.White
+import com.example.seviya.core.designsystem.components.ClientBottomBar
+import com.example.seviya.core.designsystem.components.ClientTab
+import com.example.seviya.core.designsystem.theme.AppBackground
+import com.example.seviya.core.designsystem.theme.BlueGrayText
+import com.example.seviya.core.designsystem.theme.BorderSoft
+import com.example.seviya.core.designsystem.theme.BrandBlue
+import com.example.seviya.core.designsystem.theme.BrandRed
+import com.example.seviya.core.designsystem.theme.SoftSurface
+import com.example.seviya.core.designsystem.theme.StatusCancelledBackground
+import com.example.seviya.core.designsystem.theme.StatusCancelledText
+import com.example.seviya.core.designsystem.theme.StatusCompletedBackground
+import com.example.seviya.core.designsystem.theme.StatusCompletedText
+import com.example.seviya.core.designsystem.theme.StatusConfirmedBackground
+import com.example.seviya.core.designsystem.theme.StatusConfirmedText
+import com.example.seviya.core.designsystem.theme.StatusInProgressBackground
+import com.example.seviya.core.designsystem.theme.StatusInProgressText
+import com.example.seviya.core.designsystem.theme.StatusPendingBackground
+import com.example.seviya.core.designsystem.theme.StatusPendingText
+import com.example.seviya.core.designsystem.theme.TextBluePrimary
+import com.example.seviya.core.designsystem.theme.TextOnBlueSubtitle
+import com.example.seviya.core.designsystem.theme.TextSecondary
+import com.example.seviya.core.designsystem.theme.White
 import com.example.shared.domain.entity.Appointment
 import com.example.shared.domain.entity.CancellationPolicy
 import com.example.shared.presentation.cancellation.AppointmentCancellationPreview
