+++ composeApp/src/commonMain/kotlin/com/example/seviya/UI/ClientHomeScreen.kt
@@ -42,26 +42,21 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.painter.Painter
 import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.text.SpanStyle
-import androidx.compose.ui.text.buildAnnotatedString
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.text.style.TextOverflow
-import androidx.compose.ui.text.withStyle
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.example.seviya.theme.AppBackgroundAlt
-import com.example.seviya.theme.BackgroundTopBlue
-import com.example.seviya.theme.BorderSoftAlt
-import com.example.seviya.theme.BrandBlue
-import com.example.seviya.theme.BrandBlueDeep
-import com.example.seviya.theme.BrandBlueSoft
-import com.example.seviya.theme.BrandRed
-import com.example.seviya.theme.ImageBlueSoftAlt
-import com.example.seviya.theme.TextOnBlueSubtitle
-import com.example.seviya.theme.TextPrimaryAlt
-import com.example.seviya.theme.TextSecondary
-import com.example.seviya.theme.White
+import com.example.seviya.core.designsystem.theme.AppBackgroundAlt
+import com.example.seviya.core.designsystem.theme.BackgroundTopBlue
+import com.example.seviya.core.designsystem.theme.BorderSoftAlt
+import com.example.seviya.core.designsystem.theme.BrandBlue
+import com.example.seviya.core.designsystem.theme.BrandBlueSoft
+import com.example.seviya.core.designsystem.theme.BrandRed
+import com.example.seviya.core.designsystem.theme.ImageBlueSoftAlt
+import com.example.seviya.core.designsystem.theme.TextPrimaryAlt
+import com.example.seviya.core.designsystem.theme.TextSecondary
+import com.example.seviya.core.designsystem.theme.White
 import com.example.shared.domain.entity.Category
 import com.example.shared.domain.entity.WorkerListItemData
 import com.example.shared.presentation.categories.CategoriesUiState
