+++ composeApp/src/commonMain/kotlin/com/example/seviya/UI/DashboardClientScreen.kt
@@ -39,7 +39,6 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
-import androidx.compose.ui.draw.shadow
 import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.StrokeCap
@@ -51,28 +50,28 @@ import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.example.seviya.theme.AppBackgroundAlt
-import com.example.seviya.theme.BlueGrayText
-import com.example.seviya.theme.BlueGrayTextDark
-import com.example.seviya.theme.BrandBlue
-import com.example.seviya.theme.BrandRed
-import com.example.seviya.theme.ClientHeroCardBackground
-import com.example.seviya.theme.ClientMutedDot
-import com.example.seviya.theme.ClientQuickActionBorder
-import com.example.seviya.theme.ClientQuickActionSurface
-import com.example.seviya.theme.ClientSectionCardBorder
-import com.example.seviya.theme.ClientSoftBlueRow
-import com.example.seviya.theme.ClientSoftBlueRowBorder
-import com.example.seviya.theme.DividerSoft
-import com.example.seviya.theme.Info
-import com.example.seviya.theme.Inactive
-import com.example.seviya.theme.MetricGreen
-import com.example.seviya.theme.RingTrack
-import com.example.seviya.theme.StatusCancelledText
-import com.example.seviya.theme.TextPrimaryAlt
-import com.example.seviya.theme.TextSecondary
-import com.example.seviya.theme.Warning
-import com.example.seviya.theme.White
+import com.example.seviya.core.designsystem.theme.AppBackgroundAlt
+import com.example.seviya.core.designsystem.theme.BlueGrayText
+import com.example.seviya.core.designsystem.theme.BlueGrayTextDark
+import com.example.seviya.core.designsystem.theme.BrandBlue
+import com.example.seviya.core.designsystem.theme.BrandRed
+import com.example.seviya.core.designsystem.theme.ClientHeroCardBackground
+import com.example.seviya.core.designsystem.theme.ClientMutedDot
+import com.example.seviya.core.designsystem.theme.ClientQuickActionBorder
+import com.example.seviya.core.designsystem.theme.ClientQuickActionSurface
+import com.example.seviya.core.designsystem.theme.ClientSectionCardBorder
+import com.example.seviya.core.designsystem.theme.ClientSoftBlueRow
+import com.example.seviya.core.designsystem.theme.ClientSoftBlueRowBorder
+import com.example.seviya.core.designsystem.theme.DividerSoft
+import com.example.seviya.core.designsystem.theme.Info
+import com.example.seviya.core.designsystem.theme.Inactive
+import com.example.seviya.core.designsystem.theme.MetricGreen
+import com.example.seviya.core.designsystem.theme.RingTrack
+import com.example.seviya.core.designsystem.theme.StatusCancelledText
+import com.example.seviya.core.designsystem.theme.TextPrimaryAlt
+import com.example.seviya.core.designsystem.theme.TextSecondary
+import com.example.seviya.core.designsystem.theme.Warning
+import com.example.seviya.core.designsystem.theme.White
 import com.example.shared.domain.entity.Booking
 import com.example.shared.presentation.clientDashboard.ClientDashboardUiState
 import com.example.shared.presentation.clientDashboard.ClientDashboardViewModel
