+++ composeApp/src/commonMain/kotlin/com/example/seviya/UI/DashboardWorkerScreen.kt
@@ -1,7 +1,6 @@
 package com.example.seviya.UI
 
 import androidx.compose.foundation.Canvas
-import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
 import androidx.compose.foundation.clickable
@@ -18,7 +17,6 @@ import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.navigationBarsPadding
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.statusBars
 import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.layout.windowInsetsPadding
 import androidx.compose.foundation.rememberScrollState
@@ -52,25 +50,25 @@ import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
-import com.example.seviya.theme.AccentYellow
-import com.example.seviya.theme.ActiveChipBackground
-import com.example.seviya.theme.ActiveChipText
-import com.example.seviya.theme.AppBackground
-import com.example.seviya.theme.BlueGrayText
-import com.example.seviya.theme.BlueGrayTextDark
-import com.example.seviya.theme.BlueGrayTextLight
-import com.example.seviya.theme.BorderSoft
-import com.example.seviya.theme.BrandBlue
-import com.example.seviya.theme.CardBlueDeep
-import com.example.seviya.theme.CardSurface
-import com.example.seviya.theme.MetricGreen
-import com.example.seviya.theme.OnlineGreen
-import com.example.seviya.theme.RingTrack
-import com.example.seviya.theme.SoftBlueSurface
-import com.example.seviya.theme.SoftRedSurface
-import com.example.seviya.theme.TextPrimary
-import com.example.seviya.theme.VibrantRed
-import com.example.seviya.theme.White
+import com.example.seviya.core.designsystem.theme.AccentYellow
+import com.example.seviya.core.designsystem.theme.ActiveChipBackground
+import com.example.seviya.core.designsystem.theme.ActiveChipText
+import com.example.seviya.core.designsystem.theme.AppBackground
+import com.example.seviya.core.designsystem.theme.BlueGrayText
+import com.example.seviya.core.designsystem.theme.BlueGrayTextDark
+import com.example.seviya.core.designsystem.theme.BlueGrayTextLight
+import com.example.seviya.core.designsystem.theme.BorderSoft
+import com.example.seviya.core.designsystem.theme.BrandBlue
+import com.example.seviya.core.designsystem.theme.CardBlueDeep
+import com.example.seviya.core.designsystem.theme.CardSurface
+import com.example.seviya.core.designsystem.theme.MetricGreen
+import com.example.seviya.core.designsystem.theme.OnlineGreen
+import com.example.seviya.core.designsystem.theme.RingTrack
+import com.example.seviya.core.designsystem.theme.SoftBlueSurface
+import com.example.seviya.core.designsystem.theme.SoftRedSurface
+import com.example.seviya.core.designsystem.theme.TextPrimary
+import com.example.seviya.core.designsystem.theme.VibrantRed
+import com.example.seviya.core.designsystem.theme.White
 import com.example.shared.domain.entity.Booking
 import com.example.shared.presentation.workerDashboard.WorkerDashboardUiState
 import com.example.shared.presentation.workerDashboard.WorkerDashboardViewModel
