+++ composeApp/src/commonMain/kotlin/com/example/seviya/UI/MonthlyCalendarScreen.kt
@@ -55,17 +55,16 @@ import androidx.compose.ui.graphics.graphicsLayer
 import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
-import com.example.seviya.theme.BrandBlue
-import com.example.seviya.theme.BrandRed
-import com.example.seviya.theme.InactiveSoft
-import com.example.seviya.theme.White
+import com.example.seviya.core.designsystem.theme.BrandBlue
+import com.example.seviya.core.designsystem.theme.BrandRed
+import com.example.seviya.core.designsystem.theme.InactiveSoft
+import com.example.seviya.core.designsystem.theme.White
 import com.example.shared.domain.entity.Appointment
 import com.example.shared.presentation.calendar.CalendarUserRole
 import com.example.shared.presentation.calendar.MonthlyCalendarViewModel
 import compose.icons.TablerIcons
 import compose.icons.tablericons.Apps
 import compose.icons.tablericons.ArrowLeft
-import compose.icons.tablericons.Bell
 import compose.icons.tablericons.CalendarEvent
 import compose.icons.tablericons.ChevronLeft
 import compose.icons.tablericons.ChevronRight
