+++ composeApp/src/commonMain/kotlin/com/example/seviya/UI/ProfessionalProfileScreen.kt
@@ -16,13 +16,11 @@ import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.navigationBars
 import androidx.compose.foundation.layout.offset
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.systemBarsPadding
 import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.layout.windowInsetsBottomHeight
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.foundation.shape.RoundedCornerShape
@@ -57,7 +55,30 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.example.seviya.theme.*
+import com.example.seviya.core.designsystem.theme.AppBackground
+import com.example.seviya.core.designsystem.theme.AvatarBlueSoft
+import com.example.seviya.core.designsystem.theme.BlueGrayText
+import com.example.seviya.core.designsystem.theme.BlueGrayTextDark
+import com.example.seviya.core.designsystem.theme.BlueGrayTextDarkAlt
+import com.example.seviya.core.designsystem.theme.BlueGrayTextLight
+import com.example.seviya.core.designsystem.theme.BlueGrayTextLighter
+import com.example.seviya.core.designsystem.theme.BlueGrayTextSoft
+import com.example.seviya.core.designsystem.theme.BlueTextMuted
+import com.example.seviya.core.designsystem.theme.BorderBlueLight
+import com.example.seviya.core.designsystem.theme.BorderBlueLightAlt
+import com.example.seviya.core.designsystem.theme.BorderSoft
+import com.example.seviya.core.designsystem.theme.BorderSoftAlt
+import com.example.seviya.core.designsystem.theme.BrandBlue
+import com.example.seviya.core.designsystem.theme.BrandRed
+import com.example.seviya.core.designsystem.theme.BrandYellow
+import com.example.seviya.core.designsystem.theme.DividerSoft
+import com.example.seviya.core.designsystem.theme.ImageBlueSoft
+import com.example.seviya.core.designsystem.theme.SoftBlueSurfaceAlt
+import com.example.seviya.core.designsystem.theme.SoftSurface
+import com.example.seviya.core.designsystem.theme.TextMuted
+import com.example.seviya.core.designsystem.theme.TextPrimary
+import com.example.seviya.core.designsystem.theme.TextSecondary
+import com.example.seviya.core.designsystem.theme.White
 import com.example.shared.domain.entity.Appointment
 import com.example.shared.domain.entity.CancellationPolicy
 import com.example.shared.domain.entity.PortfolioItem
