+++ composeApp/src/commonMain/kotlin/com/example/seviya/UI/RequestAppointmentScreen.kt
@@ -52,32 +52,30 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.example.seviya.theme.AppBackground
-import com.example.seviya.theme.AvatarBlueSoft
-import com.example.seviya.theme.BlueGrayText
-import com.example.seviya.theme.BlueGrayTextLight
-import com.example.seviya.theme.BlueGrayTextSoft
-import com.example.seviya.theme.BorderBlueLight
-import com.example.seviya.theme.BorderSoft
-import com.example.seviya.theme.BorderSoftAlt
-import com.example.seviya.theme.BrandBlue
-import com.example.seviya.theme.BrandRed
-import com.example.seviya.theme.CardSurface
-import com.example.seviya.theme.DividerSoft
-import com.example.seviya.theme.StatusPendingBackground
-import com.example.seviya.theme.StatusPendingText
-import com.example.seviya.theme.Success
-import com.example.seviya.theme.TextPrimary
-import com.example.seviya.theme.TextSecondary
-import com.example.seviya.theme.White
+import com.example.seviya.core.designsystem.theme.AppBackground
+import com.example.seviya.core.designsystem.theme.AvatarBlueSoft
+import com.example.seviya.core.designsystem.theme.BlueGrayText
+import com.example.seviya.core.designsystem.theme.BlueGrayTextLight
+import com.example.seviya.core.designsystem.theme.BlueGrayTextSoft
+import com.example.seviya.core.designsystem.theme.BorderBlueLight
+import com.example.seviya.core.designsystem.theme.BorderSoft
+import com.example.seviya.core.designsystem.theme.BorderSoftAlt
+import com.example.seviya.core.designsystem.theme.BrandBlue
+import com.example.seviya.core.designsystem.theme.BrandRed
+import com.example.seviya.core.designsystem.theme.CardSurface
+import com.example.seviya.core.designsystem.theme.DividerSoft
+import com.example.seviya.core.designsystem.theme.StatusPendingBackground
+import com.example.seviya.core.designsystem.theme.StatusPendingText
+import com.example.seviya.core.designsystem.theme.Success
+import com.example.seviya.core.designsystem.theme.TextPrimary
+import com.example.seviya.core.designsystem.theme.TextSecondary
+import com.example.seviya.core.designsystem.theme.White
 import com.example.shared.domain.entity.Address
 import com.example.shared.domain.entity.Appointment
 import com.example.shared.domain.entity.Service
-import com.example.shared.domain.entity.WorkerSchedule
 import com.example.shared.presentation.requestAppointment.AppointmentDayOption
 import com.example.shared.presentation.requestAppointment.AppointmentMonthOption
 import com.example.shared.presentation.requestAppointment.AppointmentTimeOption
-import com.example.shared.presentation.requestAppointment.CurrentTimeSnapshot
 import com.example.shared.presentation.requestAppointment.RequestAppointmentUiState
 import com.example.shared.presentation.requestAppointment.RequestAppointmentViewModel
 import com.example.shared.presentation.requestAppointment.REQUEST_APPOINTMENT_MONTH_SELECTION_RANGE
