+++ composeApp/src/commonMain/kotlin/com/example/seviya/UI/WorkerRequestsScreen.kt
@@ -53,7 +53,27 @@ import androidx.compose.ui.graphics.graphicsLayer
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.example.seviya.theme.*
+import com.example.seviya.core.designsystem.theme.AppBackground
+import com.example.seviya.core.designsystem.theme.BlueGrayText
+import com.example.seviya.core.designsystem.theme.BorderBlueLight
+import com.example.seviya.core.designsystem.theme.BorderUltraSoft
+import com.example.seviya.core.designsystem.theme.BrandBlue
+import com.example.seviya.core.designsystem.theme.BrandBlueDark
+import com.example.seviya.core.designsystem.theme.BrandRed
+import com.example.seviya.core.designsystem.theme.CardSurface
+import com.example.seviya.core.designsystem.theme.ClientQuickActionSurface
+import com.example.seviya.core.designsystem.theme.ClientSectionCardBorder
+import com.example.seviya.core.designsystem.theme.DividerSoft
+import com.example.seviya.core.designsystem.theme.ErrorRedSoft
+import com.example.seviya.core.designsystem.theme.ErrorText
+import com.example.seviya.core.designsystem.theme.SoftBlueSurface
+import com.example.seviya.core.designsystem.theme.SoftSurface
+import com.example.seviya.core.designsystem.theme.StatusPendingBackground
+import com.example.seviya.core.designsystem.theme.StatusPendingText
+import com.example.seviya.core.designsystem.theme.Success
+import com.example.seviya.core.designsystem.theme.TextPrimary
+import com.example.seviya.core.designsystem.theme.TextSecondary
+import com.example.seviya.core.designsystem.theme.White
 import com.example.shared.domain.entity.Appointment
 import com.example.shared.presentation.WorkerRequest.WorkerRequestsUiState
 import compose.icons.TablerIcons
