+++ composeApp/src/commonMain/kotlin/com/example/seviya/UI/WorkersListScreen.kt
@@ -49,23 +49,20 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.painter.Painter
 import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.text.SpanStyle
-import androidx.compose.ui.text.buildAnnotatedString
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.text.style.TextOverflow
-import androidx.compose.ui.text.withStyle
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.example.seviya.theme.AppBackgroundAlt
-import com.example.seviya.theme.BorderBlueLight
-import com.example.seviya.theme.BrandBlue
-import com.example.seviya.theme.BrandRed
-import com.example.seviya.theme.ImageBlueSoftAlt
-import com.example.seviya.theme.Inactive
-import com.example.seviya.theme.TextPrimaryAlt
-import com.example.seviya.theme.TextSecondary
-import com.example.seviya.theme.White
+import com.example.seviya.core.designsystem.theme.AppBackgroundAlt
+import com.example.seviya.core.designsystem.theme.BorderBlueLight
+import com.example.seviya.core.designsystem.theme.BrandBlue
+import com.example.seviya.core.designsystem.theme.BrandRed
+import com.example.seviya.core.designsystem.theme.ImageBlueSoftAlt
+import com.example.seviya.core.designsystem.theme.Inactive
+import com.example.seviya.core.designsystem.theme.TextPrimaryAlt
+import com.example.seviya.core.designsystem.theme.TextSecondary
+import com.example.seviya.core.designsystem.theme.White
 import com.example.shared.domain.entity.Address
 import com.example.shared.domain.entity.Appointment
 import com.example.shared.domain.entity.WorkZone
@@ -79,7 +76,6 @@ import compose.icons.tablericons.CalendarEvent
 import compose.icons.tablericons.ChevronRight
 import compose.icons.tablericons.Heart
 import compose.icons.tablericons.MapPin
-import compose.icons.tablericons.Moon
 import io.kamel.image.KamelImage
 import io.kamel.image.asyncPainterResource
 import kotlin.math.round
