+++ composeApp/src/commonMain/kotlin/com/example/seviya/core/designsystem/components/ClientBottomBar.kt
@@ -1,4 +1,4 @@
-package com.example.seviya.components
+package com.example.seviya.core.designsystem.components
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.*
