+++ composeApp/src/commonMain/kotlin/com/example/seviya/core/designsystem/components/FullScreenMenu.kt
@@ -1,10 +1,9 @@
-package com.example.seviya.components
+package com.example.seviya.core.designsystem.components
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
