+++ composeApp/src/commonMain/kotlin/com/example/seviya/core/designsystem/theme/Theme.kt
@@ -1,4 +1,4 @@
-package com.example.seviya.theme
+package com.example.seviya.core.designsystem.theme
 
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.lightColorScheme
