+++ composeApp/src/commonMain/kotlin/com/example/seviya/feature/categories/components/CategoriesCatalogHeader.kt
@@ -39,12 +39,12 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.withStyle
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.example.seviya.theme.AppBackgroundAlt
-import com.example.seviya.theme.BrandBlue
-import com.example.seviya.theme.BrandBlueDeep
-import com.example.seviya.theme.BrandRed
-import com.example.seviya.theme.TextOnBlueSubtitle
-import com.example.seviya.theme.White
+import com.example.seviya.core.designsystem.theme.AppBackgroundAlt
+import com.example.seviya.core.designsystem.theme.BrandBlue
+import com.example.seviya.core.designsystem.theme.BrandBlueDeep
+import com.example.seviya.core.designsystem.theme.BrandRed
+import com.example.seviya.core.designsystem.theme.TextOnBlueSubtitle
+import com.example.seviya.core.designsystem.theme.White
 import compose.icons.TablerIcons
 import compose.icons.tablericons.MapPin
 import compose.icons.tablericons.Search
