+++ composeApp/src/commonMain/kotlin/com/example/seviya/feature/categories/components/CategoriesCatalogMenu.kt
@@ -35,13 +35,13 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.compose.foundation.verticalScroll
-import com.example.seviya.theme.MenuDivider
-import com.example.seviya.theme.MenuItemSubtitle
-import com.example.seviya.theme.MenuItemTitle
-import com.example.seviya.theme.MenuSurface
-import com.example.seviya.theme.MenuTitle
-import com.example.seviya.theme.TextBluePrimary
-import com.example.seviya.theme.White
+import com.example.seviya.core.designsystem.theme.MenuDivider
+import com.example.seviya.core.designsystem.theme.MenuItemSubtitle
+import com.example.seviya.core.designsystem.theme.MenuItemTitle
+import com.example.seviya.core.designsystem.theme.MenuSurface
+import com.example.seviya.core.designsystem.theme.MenuTitle
+import com.example.seviya.core.designsystem.theme.TextBluePrimary
+import com.example.seviya.core.designsystem.theme.White
 import compose.icons.TablerIcons
 import compose.icons.tablericons.Adjustments
 import compose.icons.tablericons.CalendarEvent
