+++ composeApp/src/iosMain/kotlin/com/example/seviya/MainViewController.kt
@@ -1,6 +1,7 @@
 package com.example.seviya
 
 import androidx.compose.ui.window.ComposeUIViewController
+import com.example.seviya.app.App
 import com.example.shared.di.initKoin
 
 fun MainViewController() = ComposeUIViewController(configure = { initKoin() }) { App() }
